Cygwin: timers: use spinlock to prime hires_ns thread-safe
The current method to make hires_ns priming thread-safe isn't
thread-safe.  Rather than hoping that running the thread in
TIME_CRITICAL priority is doing the right thing, use a spinlock.

+++ winsup/cygwin/hires.h
@@ -40,7 +40,7 @@ details. */
 
 class hires_ns
 {
-  int inited;
+  LONG inited;
   LARGE_INTEGER primed_pc;
   double freq;
   void prime ();
